test(work): hold both engines to one contract; prove crash and upgrade recovery
Adds the JobEngine contract suite, run with the same bodies on the inline
engine and on DBOS: step order, retries, execution-id idempotency, dedupe
single-flight, delays, priority, lane concurrency, cancellation, evidence,
foreign versions, pruning and reset, plus the three paths DBOS broke
(submitting from a reconcile pass, from inside a step, and from an event
loop thread). Adds an e2e that boots the real app in child processes on
SQLite and PostgreSQL, kills one mid-derivative, and shows a restarted
process (or the next version) finishing the work.

Fixes found by these tests:

- artifact_derivatives.peak_rss_bytes is BIGINT: RSS in bytes overflowed
  PostgreSQL INTEGER and failed every mesh derivative there;
- startup forgets queued-pass marks a dead process left, so recovery is
  not held back for the submit grace window;
- the heartbeat thread runs in the starter's context;
- import cycles removed: the session admission gate moves to app.db and the
  dispatch status to the printing dispatch definition;
- adds the bootstrap.work suite (roles, topology validation, start, stop,
  restore) and tests for the database admission gate.

+++ backend/alembic/versions/a83ea973fad4_job_engine_schema_confirmed_data_.py
@@ -227,7 +227,7 @@ def upgrade() -> None:
     sa.Column('storage_key', sa.String(length=2048), nullable=True),
     sa.Column('output_json', sa.Text(), nullable=False),
     sa.Column('duration_ms', sa.Integer(), nullable=True),
-    sa.Column('peak_rss_bytes', sa.Integer(), nullable=True),
+    sa.Column('peak_rss_bytes', sa.BigInteger(), nullable=True),
     sa.Column('created_at', sa.DateTime(), nullable=False),
     sa.Column('updated_at', sa.DateTime(), nullable=False),
     sa.ForeignKeyConstraint(['file_id'], ['files.id'], name=op.f('fk_artifact_derivatives_file_id_files'), ondelete='CASCADE'),
